Update minify.php
The code results in a syntax error in the CSS style sheet output because media queries targeting a screen device doesn't have a space between the 'and' keyword and the opening parenthesis of media features. For a quick fix, the following pattern '/screen and/' and the replacement string '$0 ' added to the preg_replace() function (at Line 198).
Additionally, parameters are added to the function to remove leading zeros from the beginning of a number before the decimal point in a string.